General cleanup
In C the way to designate a function or field "private" is to mark it
static. This tells the compiler that the function/field is not needed
outside this compilation unit (usually one file). This also means the
linker doesn't need to include space in the symbol table for these
static functions/fields thereby reducing app size.

For example, on basalt marking the appropriate things static reduces the
app size from 27464 bytes to 27296 bytes. That's 168 bytes, which isn't
a lot, but if you're ever constrained for memory every byte counts. For
some reason diorite goes from 27464 to 25543; I'm not going to complain.

In addition, for functions that take no arguments, specify void in the
argument list. C allows functions without any arguments defined to
actually be called with any number of arguments. This can corrupt the
stack space and lead to bad things. Better to just head this off by
saying, "This function takes absolutely no arguments."